requirements.txt
pandas
pinecone
python-dotenv
sentence-transformers
FlagEmbedding
fastapi
google-generativeai
uvicorn
